==> analysis/02_fig2_mrg_inclusion_overlap.Rmd <==
---
title: "Medical Gene Benchmark Inclusion and Overlap"
date: '`r Sys.Date()`'
output: 
    bookdown::html_document2:
        toc: true
        toc_float: true
        df_print: paged
        code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(here)
library(readxl)
library(ggpubr)
```

## Loading and Tidying
# Loading and Tidying Data
```{r}
## Gene source and testing modality information
mrg_info_tbl <- here("data", "tidy", "gene_list_info.tsv.gz") %>%
    read_tsv(col_types = "cccccddddddd")

## Gene Stratfication Overlap - no longer needed
# gene_strat_df <- read_tsv(here("data","tidy","gene_strat_overlap.tsv"),
#                           col_types = cols())


## Benchmark Included Bases
bench_cov_df <- here("data","tidy","bench_coverage.tsv.gz") %>% 
    read_tsv(col_types = "ccccccdil")

## Alll Difficult Strat Overlap
diff_gene_overlap_df <- here("data", "tidy", "strat_overlap.tsv.gz") %>%
    read_tsv(col_types = "cccciiciiid") %>%
    filter(strat == "allDiff", region == "gene") %>%
    rename(gene_size = region_size) %>%
    select(-region)

## TODO - replace with new updated coverage data
## Benchmark and Asm Coverage information 
# bench_asm_cov_df <- here("data","tidy","gene_asm_bench_coverage.tsv") %>% 
#     read_tsv(col_types = "ccclccddd")

## Benchmark Variant tbl
mrg_var_df <- read_tsv(here("data","tidy","var_tbl.tsv.gz"),
                       col_types = "ccccciccccici") %>% 
    filter(benchmark == "mrg",
           ref == "GRCh38")
```

Getting gene categories
```{r}
mrg_gene_cats <- bench_cov_df %>% 
    filter(str_detect(coords, "[MTXY]",negate = TRUE),
           ref == "GRCh38",
           region == "gene",
           (benchmark == "v4" & bench_type == "smallvar") | 
           (benchmark == "mrg" & bench_type == "union"),
           ) %>% 
     select(-bench_type) %>% 
    pivot_wider(names_from = benchmark, values_from = cov) %>%
    mutate(gene_cat = case_when(mrg > 0.9 ~ "CMRG", #in_mrg ~ "MRG",
                                v4 > 0.9 ~ "V4",
                                v4 < 0.9 ~ "Excluded",
                                TRUE ~ "???"))
```

## Panels


### MRG Gene Coverage
```{r}
mrg_gene_cov <- bench_cov_df %>%
    filter(in_mrg == TRUE,
            ref == "GRCh38",
           str_detect(coords, "[MTXY]",negate = TRUE),
           benchmark == "mrg",
           bench_type == "union",
           region != "intron") %>% 
    mutate(region = if_else(region == "gene", "Gene", "Exon"))

(mrg_gene_cov_plt <- ggplot(mrg_gene_cov) + 
    geom_histogram(aes(x = cov, fill = region), 
                   position = "dodge",
                   color = "grey40",
                   size = 0.25, bins = 25) + 
    theme_bw() +
    labs(x = "Fraction included in\nCMRG benchmark", 
         y = "Count", 
         fill = "") +
    theme(legend.position = c(0.25,0.8), 
          legend.direction = "vertical", 
          legend.background = element_blank()))
```

### Difficult Gene Overlap
```{r}
mrg_gene_cats %>% 
    select(gene, ref, coords, gene_cat) %>% 
    count(ref, gene_cat)
```

```{r}
diff_overlap_df <- mrg_gene_cats %>% 
    select(gene, ref, coords, gene_cat) %>% 
    left_join(diff_gene_overlap_df) %>%
    filter(
        ref == "GRCh38",
        gene_cat %in% c("V4", "CMRG", "Excluded"),
        str_detect(coords, "[MTXY]", negate = TRUE),
        strat == "allDiff"
    ) %>%
    mutate(gene_cat = factor(
        gene_cat,
        level = c("Excluded", "V4", "CMRG"),
        labels = c("Excluded",
                   ">90% included in V4.2.1",
                   ">90% included in CMRG Bench")
    ))

cum_rank_df <- diff_overlap_df  %>% 
    group_by(gene_cat) %>% 
    arrange(overlap) %>% 
    mutate(gene_rank = 1:n(),
           n_genes = n(),
           gene_rank_frac = (n_genes - gene_rank)/n_genes)

cum_rank_text_df <- cum_rank_df %>% 
    mutate(overlap_round = round(overlap,2)) %>% 
    filter(overlap_round %in% c(0.20,1)) %>% 
    group_by(gene_cat, overlap_round) %>% 
    top_n(1, gene_rank_frac) %>% 
    mutate(n_genes_gt15_overlap = n_genes - gene_rank)
```



```{r}
(
    diff_overlap_plt <- ggplot(cum_rank_df) +
        geom_segment(
            data = cum_rank_text_df,
            aes(
                x = overlap_round,
                xend = overlap_round,
                y = 0,
                yend = gene_rank_frac * 100
            ),
            linetype = 2
        ) +
        geom_ribbon(
            aes(
                ymin = 0,
                ymax = gene_rank_frac * 100,
                x = overlap,
                fill = gene_cat
            ),
            alpha = 0.25
        ) +
        geom_line(aes(
            y = gene_rank_frac * 100, x = overlap, color = gene_cat
        )) +
        geom_point(
            data = cum_rank_text_df,
            aes(x = overlap, y = gene_rank_frac * 100, fill = gene_cat),
            shape = 21
        ) +
        geom_text(
            data = cum_rank_text_df,
            aes(x = overlap,
                y = gene_rank_frac * 100,
                label = n_genes_gt15_overlap),
            nudge_x = -0.01,
            vjust = "bottom",
            hjust = "right"
        ) +
        scale_fill_brewer(type = "qual", palette = 2) +
        scale_color_brewer(type = "qual", palette = 2) +
        theme_bw() +
        theme(
            legend.position = c(0.70, 0.85),
            legend.background = element_blank()
        ) +
        labs(
            x = "Fraction of gene region with\nchallenging sequence or variants",
            y = "% of Genes with > X Challenging",
            color = "",
            fill = ""
        )
)
```


### Small Variant Size Distribution
```{r}
(smv_size_dist_plt <- mrg_var_df %>% 
    filter(TYPE == "INDEL", bench_type == "smallvar") %>% 
    mutate(gene_part = if_else(EXON == ".", "Intron","Exon"),
           gene_part = factor(gene_part, levels = c("Intron","Exon"))) %>% 
    ggplot() + 
    geom_bar(aes(x = var_size, fill = gene_part),
             color = "grey40", size = 0.25) + 
        scale_fill_brewer(type = "qual", palette = 3) +
    theme_bw() +
    labs(x = "Small INDEL Size (bp)", y = "# of Variants", fill = "Gene Part") + 
    theme(legend.position = c(0.15, 0.8), legend.background = element_blank()))
```

### Large Variant Size Distribution 


```{r}
breaks <- c(50, 100, 500, 1000, 5000, 10000)
log_breaks <- log10(breaks)

sv_log_breaks <- sort(c(-log_breaks + log10(25), 0,  log_breaks - log10(25)))

sv_breaks <- sort(c(-breaks, 0, breaks))

(sv_size_dist_plt <- mrg_var_df %>% 
    filter(TYPE == "INDEL", bench_type == "SV") %>% 
    mutate(gene_part = if_else(EXON == ".", "Intron","Exon"),
           gene_part = factor(gene_part, levels = c("Intron","Exon"))) %>%         
    ## Log scale var length with 1 offset 
    mutate(var_size = if_else(var_size < 0, 
                              -log10(abs(var_size)) + log10(25),
                              log10(var_size) - log10(25))) %>% 
    ggplot() + 
    geom_histogram(aes(x = var_size, fill = gene_part), 
                   color = "grey40", size = 0.25, binwidth = 0.125) +
    scale_x_continuous(breaks = sv_log_breaks,
                       labels = sv_breaks) + 
        scale_fill_brewer(type = "qual", palette = 3) + 
    theme_bw() +
    theme(axis.text.x = element_text(angle = -45, hjust = 0),
          legend.position = c(0.15, 0.8), legend.background = element_blank()) + 
    labs(x = "Large SV Size (bp)", y = "# of Variants"))
```


# Combined Figure
```{r}
(
    combined_fig <- ggarrange(
        ggarrange(
            mrg_gene_cov_plt,
            diff_overlap_plt,
            nrow = 1,
            widths = c(1,1),
            labels = "AUTO"),
        ggarrange(
            smv_size_dist_plt,
            sv_size_dist_plt + rremove(object = "legend"),
            nrow = 1,
            labels = c("C", "D")
        ),
        nrow = 2,
        heights = c(2,1)
    )
)
```


```{r}
ggsave(filename = here("figures", "fig2_mrg_inclusion_overlap.png"),
       plot = combined_fig, width = 8, height = 8, dpi = 300)
```




# Session Information
## System Information
```{r}
sessioninfo::platform_info()
```


## Package Versions
```{r}
sessioninfo::package_info() %>% 
    filter(attached = TRUE) %>% 
    select(package, loadedversion, date, source) %>%
    knitr::kable(booktabs = TRUE, row.names = FALSE)
```
